requirements.txt
streamlit
fredapi
plotly
pandas
python-dateutil
requests
